requirements.txt
yfinance
pandas
python-telegram-bot
flask